+++ mitra_models/src/markers/types.rs
@@ -7,7 +7,7 @@ use crate::database::{
     DatabaseTypeError,
 };
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub enum Timeline {
     Home,
     Notifications,