+++ src/mastodon_api/markers/types.rs
@@ -6,18 +6,21 @@ use mitra_validators::errors::ValidationError;
 
 #[derive(Deserialize)]
 pub struct MarkerQueryParams {
-    #[serde(rename = "timeline[]")]
-    pub timeline: String,
+    pub timeline: Vec<String>,
 }
 
 impl MarkerQueryParams {
-    pub fn to_timeline(&self) -> Result<Timeline, ValidationError> {
-        let timeline = match self.timeline.as_ref() {
-            "home" => Timeline::Home,
-            "notifications" => Timeline::Notifications,
-            _ => return Err(ValidationError("invalid timeline name")),
+    pub fn to_timelines(&self) -> Result<Vec<Timeline>, ValidationError> {
+        let mut timelines = vec![];
+        for value in &self.timeline {
+            let timeline = match value.as_str() {
+                "home" => Timeline::Home,
+                "notifications" => Timeline::Notifications,
+                _ => return Err(ValidationError("invalid timeline name")),
+            };
+            timelines.push(timeline);
         };
-        Ok(timeline)
+        Ok(timelines)
     }
 }