+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Log object ID if poll data is not valid.
+
 ## [3.13.1] - 2025-01-03
 
 ### Fixed