+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Allow to run separate worker for processing of incoming activity queue.
+
 ### Changed
 
 - Log object ID if poll data is not valid.