+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Log object ID if poll data is not valid.
 - Log database errors coming from inbox handler with level ERROR.
 - Log database connection pool size.
+- Release database connection before delivering activity.
 
 ### Fixed