+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.14.0] - 2025-01-14
+
 ### Added
 
 - Allow to run separate worker for processing of incoming activity queue.